move UnsupportedStatistic inside Statistic.ts, add static create()
